pyproject.toml
[project]
name = "anti-russian-spam-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipdb>=0.13.13",
    "pyright>=1.1.408",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]

[project.scripts]
anti-spam-bot = "main:main"
